requirements.txt
# Core
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
datasets>=2.18.0
huggingface-hub>=0.21.0

# Numerical
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
einops>=0.7.0

# Notebook
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0
